Finished.
The file accepts three variables: infile, outfile, terms.

The infile is the file that contains the text to be searched.
The outfile is the file that will contain the results of the search.
The terms is a list of terms to be searched for.

The file will search for the terms in the infile and output the results to the outfile.

The file will also output the line on which each was found.

The infile defaults to "origin.txt" and the outfile defaults to "origin_output.txt".
The terms default to herit.

The file is correctly accepting alternate infile, outfile, and terms.